% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-create-org.R
\name{admin_create_org}
\alias{admin_create_org}
\title{WIP - Create an organization}
\usage{
admin_create_org(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create an organization
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
