% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-create-repo.R
\name{admin_create_repo}
\alias{admin_create_repo}
\title{WIP - Create a repository on behalf a user}
\usage{
admin_create_repo(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create a repository on behalf a user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
