% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-delete-user.R
\name{admin_delete_user}
\alias{admin_delete_user}
\title{WIP - Delete a user}
\usage{
admin_delete_user(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Delete a user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
