% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-edit-user.R
\name{admin_edit_user}
\alias{admin_edit_user}
\title{WIP - Edit an existing user}
\usage{
admin_edit_user(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Edit an existing user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
