% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-current-user-repo.R
\name{create_current_user_repo}
\alias{create_current_user_repo}
\title{WIP - Create a repository}
\usage{
create_current_user_repo(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
