% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-fork.R
\name{create_fork}
\alias{create_fork}
\title{WIP - Fork a repository}
\usage{
create_fork(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Fork a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
