% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-current-user-repo.R
\name{create_user_repo}
\alias{create_user_repo}
\title{Create a repository as the owner}
\usage{
create_user_repo(name, description = NULL, gitignores = NULL,
  readme = "Default", license = NULL, private = FALSE, auto_init = TRUE,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{name}{Repository name. Mandatory. A good repository name is composed of short, memorable, and unique keywords.}

\item{description}{Repository description. Optional.}

\item{gitignores}{Use a \code{.gitignore} template. Supply a valid name you see in the dropdown. Case matters. Optional.}

\item{readme}{use the \code{Default} empty \code{README.md}}

\item{license}{Use a content license template. Supply a valid name you see in the dropdown. Case matters. Optional.}

\item{private}{logical. Defaults to \code{FALSE}}

\item{auto_init}{auto-initialize the repo? Default is \code{TRUE}}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
list (invisibly) with the status result of the API
}
\description{
Create a repository as the owner
}
\examples{
\dontrun{
}
}
