% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumpets-package.R
\docType{package}
\name{crumpets}
\alias{crumpets}
\alias{crumpets-package}
\title{Tools to Work with and Manage 'gitea'}
\description{
The 'gitea' project is a 'GitHub'-/'GitLab'-like web user interface
to 'git' repositories. Methods are provided to access and administer 'gitea'
programmatically.
}
\author{
Bob Rudis (bob@rud.is)
}
