% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-add-time.R
\name{issue_add_time}
\alias{issue_add_time}
\title{WIP - Add a tracked time to a issue}
\usage{
issue_add_time(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Add a tracked time to a issue
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
