% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-clear-labels.R
\name{issue_clear_labels}
\alias{issue_clear_labels}
\title{WIP - Remove all labels from an issue}
\usage{
issue_clear_labels(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Remove all labels from an issue
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
