% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-create-comment.R
\name{issue_create_comment}
\alias{issue_create_comment}
\title{WIP - Add a comment to an issue}
\usage{
issue_create_comment(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Add a comment to an issue
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
