% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-create-issue.R
\name{issue_create_issue}
\alias{issue_create_issue}
\title{WIP - Create an issue}
\usage{
issue_create_issue(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create an issue
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
