% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-create-milestone.R
\name{issue_create_milestone}
\alias{issue_create_milestone}
\title{WIP - Create a milestone}
\usage{
issue_create_milestone(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create a milestone
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
