% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-delete-label.R
\name{issue_delete_label}
\alias{issue_delete_label}
\title{WIP - Delete a label}
\usage{
issue_delete_label(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Delete a label
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
