% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-edit-label.R
\name{issue_edit_label}
\alias{issue_edit_label}
\title{WIP - Update a label}
\usage{
issue_edit_label(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Update a label
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
