% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-get-comments.R
\name{issue_get_comments}
\alias{issue_get_comments}
\title{WIP - List all comments on an issue}
\usage{
issue_get_comments(owner, repo, id, string,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{id}{index of the issue}

\item{string}{if provided, only comments updated since the specified time are returned.}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List all comments on an issue
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
