% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-get-labels.R
\name{issue_get_labels}
\alias{issue_get_labels}
\title{WIP - Get an issue's labels}
\usage{
issue_get_labels(owner, repo, index,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{index}{index of the issue}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get an issue's labels
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
