% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-get-repo-comments.R
\name{issue_get_repo_comments}
\alias{issue_get_repo_comments}
\title{WIP - List all comments in a repository}
\usage{
issue_get_repo_comments(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List all comments in a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
