% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-list-issues.R
\name{issue_list_issues}
\alias{issue_list_issues}
\title{WIP - List a repository's issues}
\usage{
issue_list_issues(owner, repo, state, page,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{state}{whether issue is open or closed}

\item{page}{page number of requested issues}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List a repository's issues
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
