% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-remove-label.R
\name{issue_remove_label}
\alias{issue_remove_label}
\title{WIP - Remove a label from an issue}
\usage{
issue_remove_label(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Remove a label from an issue
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
