% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-replace-labels.R
\name{issue_replace_labels}
\alias{issue_replace_labels}
\title{WIP - Replace an issue's labels}
\usage{
issue_replace_labels(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Replace an issue's labels
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
