% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue-tracked-times.R
\name{issue_tracked_times}
\alias{issue_tracked_times}
\title{WIP - List an issue's tracked times}
\usage{
issue_tracked_times(owner, repo, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List an issue's tracked times
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
