% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-forks.R
\name{list_forks}
\alias{list_forks}
\title{WIP - List a repository's forks}
\usage{
list_forks(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List a repository's forks
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
