% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-add-team-member.R
\name{org_add_team_member}
\alias{org_add_team_member}
\title{WIP - Remove a repository from a team}
\usage{
org_add_team_member(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Remove a repository from a team
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
