% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-delete-hook.R
\name{org_delete_hook}
\alias{org_delete_hook}
\title{WIP - Delete a hook}
\usage{
org_delete_hook(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Delete a hook
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
