% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-delete-member.R
\name{org_delete_member}
\alias{org_delete_member}
\title{WIP - Remove a member from an organization}
\usage{
org_delete_member(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Remove a member from an organization
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
