% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-edit.R
\name{org_edit}
\alias{org_edit}
\title{WIP - Edit an organization}
\usage{
org_edit(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Edit an organization
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
