% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-edit-hook.R
\name{org_edit_hook}
\alias{org_edit_hook}
\title{WIP - Update a hook}
\usage{
org_edit_hook(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Update a hook
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
