% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-get.R
\name{org_get}
\alias{org_get}
\title{Get an organization}
\usage{
org_get(org, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{org}{name of the organization to get}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
\code{list}
}
\description{
Get an organization
}
\examples{
\dontrun{
org_get("hrbrville")
}
}
