% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-get-team.R
\name{org_get_team}
\alias{org_get_team}
\title{WIP - Get a team}
\usage{
org_get_team(id, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{id}{id of the team to get}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get a team
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
