% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-list-repos.R
\name{org_list_repos}
\alias{org_list_repos}
\title{WIP - List an organization's repos}
\usage{
org_list_repos(org, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{org}{name of the organization}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List an organization's repos
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
