% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-list-user-orgs.R
\name{org_list_user_orgs}
\alias{org_list_user_orgs}
\title{WIP - List a user's organizations}
\usage{
org_list_user_orgs(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List a user's organizations
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
