% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-publicize-member.R
\name{org_publicize_member}
\alias{org_publicize_member}
\title{WIP - Publicize a user's membership}
\usage{
org_publicize_member(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Publicize a user's membership
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
