% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-markdown-raw.R
\name{render_markdown_raw}
\alias{render_markdown_raw}
\title{WIP - Render raw markdown as HTML}
\usage{
render_markdown_raw(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Render raw markdown as HTML
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
