% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-add-collaborator.R
\name{repo_add_collaborator}
\alias{repo_add_collaborator}
\title{WIP - Add a collaborator to a repository}
\usage{
repo_add_collaborator(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Add a collaborator to a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
