% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-check-collaborator.R
\name{repo_check_collaborator}
\alias{repo_check_collaborator}
\title{WIP - Check if a user is a collaborator of a repository}
\usage{
repo_check_collaborator(owner, repo, collaborator,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{collaborator}{username of the collaborator}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Check if a user is a collaborator of a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
