% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-create-hook.R
\name{repo_create_hook}
\alias{repo_create_hook}
\title{WIP - Create a hook}
\usage{
repo_create_hook(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create a hook
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
