% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-create-release.R
\name{repo_create_release}
\alias{repo_create_release}
\title{WIP - Create a release}
\usage{
repo_create_release(owner, repo, body,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{body}{NA}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create a release
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
