% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-delete-release.R
\name{repo_delete_release}
\alias{repo_delete_release}
\title{WIP - Delete a release}
\usage{
repo_delete_release(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Delete a release
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
