% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-edit-hook.R
\name{repo_edit_hook}
\alias{repo_edit_hook}
\title{WIP - Edit a hook in a repository}
\usage{
repo_edit_hook(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Edit a hook in a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
