% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get.R
\name{repo_get}
\alias{repo_get}
\title{Get a repository}
\usage{
repo_get(owner, repo, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
Get a repository
}
\examples{
\dontrun{
}
}
