% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get-by-id.R
\name{repo_get_by_id}
\alias{repo_get_by_id}
\title{WIP - Get a repository by id}
\usage{
repo_get_by_id(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get a repository by id
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
