% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get-editor-config.R
\name{repo_get_editor_config}
\alias{repo_get_editor_config}
\title{WIP - Get the EditorConfig definitions of a file in a repository}
\usage{
repo_get_editor_config(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get the EditorConfig definitions of a file in a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
