% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get-pull-request.R
\name{repo_get_pull_request}
\alias{repo_get_pull_request}
\title{WIP - Get a pull request}
\usage{
repo_get_pull_request(owner, repo, index,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{index}{index of the pull request to get}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get a pull request
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
