% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get-raw-file.R
\name{repo_get_raw_file}
\alias{repo_get_raw_file}
\title{Get a file from a repository}
\usage{
repo_get_raw_file(owner, repo, filepath,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{filepath}{filepath of the file to get}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
if MIME type was intuited and \code{\link[httr:content]{httr::content()}} can decode it, then a decoded
object, otherwise a raw vector
}
\description{
Get a file from a repository
}
\examples{
\dontrun{
repo_get_raw_file("hrbrmstr", "ggalt", "/man/figures/coordproj01.png")
repo_get_raw_file("hrbrmstr", "securitytxt", "/R/RcppExports.R")
}
}
