% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-list-hooks.R
\name{repo_list_hooks}
\alias{repo_list_hooks}
\title{WIP - List the hooks in a repository}
\usage{
repo_list_hooks(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the hooks in a repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
