% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-list-subscribers.R
\name{repo_list_subscribers}
\alias{repo_list_subscribers}
\title{WIP - List a repo's watchers}
\usage{
repo_list_subscribers(owner, repo,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List a repo's watchers
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
