% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-pull-request-is-merged.R
\name{repo_pull_request_is_merged}
\alias{repo_pull_request_is_merged}
\title{WIP - Check if a pull request has been merged}
\usage{
repo_pull_request_is_merged(owner, repo, index,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{index}{index of the pull request}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Check if a pull request has been merged
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
