% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-add-email.R
\name{user_add_email}
\alias{user_add_email}
\title{WIP - Add email addresses}
\usage{
user_add_email(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Add email addresses
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
