% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-check-starring.R
\name{user_current_check_starring}
\alias{user_current_check_starring}
\title{WIP - Whether the authenticated is starring the repo}
\usage{
user_current_check_starring(owner, repo,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Whether the authenticated is starring the repo
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
