% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-delete-follow.R
\name{user_current_delete_follow}
\alias{user_current_delete_follow}
\title{WIP - Unfollow a user}
\usage{
user_current_delete_follow(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Unfollow a user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
