% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-delete-star.R
\name{user_current_delete_star}
\alias{user_current_delete_star}
\title{WIP - Unstar the given repo}
\usage{
user_current_delete_star(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Unstar the given repo
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
