% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-delete-subscription.R
\name{user_current_delete_subscription}
\alias{user_current_delete_subscription}
\title{WIP - Unwatch a repo}
\usage{
user_current_delete_subscription(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Unwatch a repo
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
