% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-list-following.R
\name{user_current_list_following}
\alias{user_current_list_following}
\title{WIP - List the users that the authenticated user is following}
\usage{
user_current_list_following(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the users that the authenticated user is following
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
