% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-list-keys.R
\name{user_current_list_keys}
\alias{user_current_list_keys}
\title{WIP - List the authenticated user's public keys}
\usage{
user_current_list_keys(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the authenticated user's public keys
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
