% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-list-subscriptions.R
\name{user_current_list_subscriptions}
\alias{user_current_list_subscriptions}
\title{WIP - List repositories watched by the authenticated user}
\usage{
user_current_list_subscriptions(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List repositories watched by the authenticated user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
