% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-post-key.R
\name{user_current_post_key}
\alias{user_current_post_key}
\title{WIP - Create a public key}
\usage{
user_current_post_key(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Create a public key
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
