% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-tracked-times.R
\name{user_current_tracked_times}
\alias{user_current_tracked_times}
\title{WIP - List the current user's tracked times}
\usage{
user_current_tracked_times(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the current user's tracked times
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
