% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-get-current.R
\name{user_get_current}
\alias{user_get_current}
\title{Get the authenticated user}
\usage{
user_get_current(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
\code{list}
}
\description{
Get the authenticated user
}
\examples{
\dontrun{
user_get_current()
}
}
