% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-list-emails.R
\name{user_list_emails}
\alias{user_list_emails}
\title{WIP - List the authenticated user's email addresses}
\usage{
user_list_emails(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the authenticated user's email addresses
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
