% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-list-gpg-keys.R
\name{user_list_gpg_keys}
\alias{user_list_gpg_keys}
\title{WIP - List the given user's GPG keys}
\usage{
user_list_gpg_keys(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the given user's GPG keys
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
