% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-list-keys.R
\name{user_list_keys}
\alias{user_list_keys}
\title{WIP - List the given user's public keys}
\usage{
user_list_keys(username, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{username}{username of user}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the given user's public keys
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
