% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-list-starred.R
\name{user_list_starred}
\alias{user_list_starred}
\title{WIP - The repos that the given user has starred}
\usage{
user_list_starred(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - The repos that the given user has starred
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
