% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-search.R
\name{user_search}
\alias{user_search}
\title{WIP - Search for users}
\usage{
user_search(q, limit, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{q}{keyword}

\item{limit}{maximum number of users to return}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Search for users
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
