% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-is-member.R
\name{org_is_member}
\alias{org_is_member}
\title{WIP - Check if a user is a member of an organization}
\usage{
org_is_member(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Check if a user is a member of an organization
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
