% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-list-team-members.R
\name{org_list_team_members}
\alias{org_list_team_members}
\title{WIP - List a team's members}
\usage{
org_list_team_members(id, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{id}{id of the team}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List a team's members
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
