% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-list-user-orgs.R
\name{org_list_user_orgs}
\alias{org_list_user_orgs}
\title{List a user's organizations}
\usage{
org_list_user_orgs(username, api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{username}{username of user}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
List a user's organizations
}
\note{
this appears to not be working on the Gitea side
}
\examples{
\dontrun{
org_list_user_orgs("hrbrmstr")
}
}
