% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get-combined-status-by-ref.R
\name{repo_get_combined_status_by_ref}
\alias{repo_get_combined_status_by_ref}
\title{WIP - Get a commit's combined status, by branch/tag/commit reference}
\usage{
repo_get_combined_status_by_ref(owner, repo, ref,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{ref}{name of branch/tag/commit}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get a commit's combined status, by branch/tag/commit reference
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
