% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-get-hook.R
\name{repo_get_hook}
\alias{repo_get_hook}
\title{WIP - Get a hook}
\usage{
repo_get_hook(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get a hook
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
