% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-migrate.R
\name{repo_migrate}
\alias{repo_migrate}
\title{WIP - Migrate a remote git repository}
\usage{
repo_migrate(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Migrate a remote git repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
