% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo-mirror-sync.R
\name{repo_mirror_sync}
\alias{repo_mirror_sync}
\title{WIP - Sync a mirrored repository}
\usage{
repo_mirror_sync(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Sync a mirrored repository
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
