% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-check-following.R
\name{user_current_check_following}
\alias{user_current_check_following}
\title{WIP - Check whether a user is followed by the authenticated user}
\usage{
user_current_check_following(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Check whether a user is followed by the authenticated user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
