% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-current-get-gpg-key.R
\name{user_current_get_gpg_key}
\alias{user_current_get_gpg_key}
\title{WIP - Get a GPG key}
\usage{
user_current_get_gpg_key(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - Get a GPG key
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
