% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-list-repos.R
\name{user_list_repos}
\alias{user_list_repos}
\title{WIP - List the repos owned by the given user}
\usage{
user_list_repos(api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{api_endpoint}{URL prefix for your gitea server}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List the repos owned by the given user
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
