% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-tracked-times.R
\name{user_tracked_times}
\alias{user_tracked_times}
\title{WIP - List a user's tracked times in a repo}
\usage{
user_tracked_times(owner, repo, user,
  api_endpoint = Sys.getenv("GITEA_BASE_URL"),
  gitea_token = Sys.getenv("GITEA_PAT"))
}
\arguments{
\item{owner}{owner of the repo}

\item{repo}{name of the repo}

\item{user}{username of user}

\item{api_endpoint}{URL prefix for your gitea server (no trailing '/')}

\item{gitea_token}{NOTE: we use \code{access_token} in the package}
}
\value{
something
}
\description{
WIP - List a user's tracked times in a repo
}
\note{
NOT IMPLEMENTED YET
}
\examples{
\dontrun{
}
}
